requirements.txt
mglearn==0.2.0
numpy>=1.22.3
matplotlib>=3.7.0
scipy>=1.10.0
scikit-learn>=1.2.1
pandas>=1.5.3
opencv-python>=4.5.5.64
datasets==2.16.1
transformers[torch]==4.37.2
evaluate==0.4.1
timm==0.9.12
albumentations==1.3.1
pycocotools==2.0.7
torchvision~=0.17.2
tqdm~=4.66.4
pillow~=10.3.0
ultralytics~=8.2.42
